Split structure assembly logic into another module
